move unit test files into correct directory
