+++ src/applications/simple-forms/21-4142/tests/unit/pages/preparerAddress1.unit.spec.jsx
@@ -1,13 +1,13 @@
 import {
   testNumberOfErrorsOnSubmit,
   testNumberOfFields,
-} from '../../../shared/tests/pages/pageTests.spec';
+} from '../../../../shared/tests/pages/pageTests.spec';
 import {
   preparerIdentificationFields,
   veteranFields,
   veteranIsSelfText,
-} from '../../definitions/constants';
-import formConfig from '../../config/form';
+} from '../../../definitions/constants';
+import formConfig from '../../../config/form';
 
 const {
   schema,
