+++ src/applications/simple-forms/21-4142/tests/unit/pages/recordsRequested.unit.spec.jsx
@@ -1,9 +1,9 @@
 import {
   testNumberOfErrorsOnSubmit,
   testNumberOfFields,
-} from '../../../shared/tests/pages/pageTests.spec';
-import { patientIdentificationFields } from '../../definitions/constants';
-import formConfig from '../../config/form';
+} from '../../../../shared/tests/pages/pageTests.spec';
+import { patientIdentificationFields } from '../../../definitions/constants';
+import formConfig from '../../../config/form';
 
 const {
   schema,
